move validate_arguments use case under programs/ and add ValidateArgumentsUseCase for the view
